+++ resources/js/components/app-header.tsx
@@ -1,5 +1,5 @@
 import { Breadcrumbs } from '@/components/breadcrumbs';
-import FormNewBookmark from '@/components/form-new-bookmark';
+import FormNewBookmark from '@/components/forms/form-new-bookmark';
 import { Icon } from '@/components/icon';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { Button } from '@/components/ui/button';
